hslt: squashed patches into one commit.
the git history was just a little messy and added and removed sizable assets
which i don't want to clutter the git history and take up space.

the code is also not very polished yet, we'll continue to improve it.
this commit is provided as an effort for complete disclosure and to spark
new research, it may require some work to get nice pictures out of it.